requirements.txt
google-genai>=1.30.0
PyMuPDF>=1.24.0
PyYAML>=6.0

